agent-framework-app/agent-framework-deepagent/agent-workflow-builder/backend/requirements.txt
# Backend Dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.1.0
sqlmodel==0.0.22
alembic==1.13.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
aiofiles==23.2.1

# Microsoft Agent Framework - CORRECTED
agent-framework

# Azure integrations
azure-identity==1.19.0
azure-keyvault-secrets==4.7.0
azure-ai-openai

# MCP Protocol support
modelcontextprotocol

# OpenTelemetry for observability (included with Agent Framework)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
black==24.10.0
ruff==0.7.4
mypy==1.7.1

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0